Update script version check

+++ scripts/install/armadillo-check-update.sh
@@ -1,21 +1,17 @@
 #!/bin/bash
 
+ARMADILLO_UPDATESCR_VERSION=0.0.1
 
-ARMADILLO_URL=https://github.com/molgenis/molgenis-service-armadillo/
-ARMADILLO_PROFILE=default
-ARMADILLO_PATH=/usr/share/armadillo
-ARMADILLO_CFG_PATH=/etc/armadillo
-ARMADILLO_SYS_USER=armadillo
-ARMADILLO_LOG_PATH=/var/log/armadillo
-ARMADILLO_AUDITLOG=$ARMADILLO_LOG_PATH/audit.log
-ARMADILLO_DATADIR=$ARMADILLO_PATH/data
+check_self_update() {
+    
+   
+}
 
 check_armadillo_update() {
 
   FILE=$(readlink "$ARMADILLO_PATH/application/armadillo.jar")
   VERSION_USED=$(curl -L -s -H 'Accept: application/json' -s http://localhost:8080/actuator/info | sed -e 's/.*"version":"\([^"]*\)".*/\1/')
-
-
+    
   LATEST_RELEASE=$(curl -L -s -H 'Accept: application/json' -s $ARMADILLO_URL/releases/latest)
   ARMADILLO_LATEST_VERSION=$(echo "$LATEST_RELEASE" | sed -e 's/.*"tag_name":"\([^"]*\)".*/\1/' | sed -e s/.*armadillo-service-//)
 
@@ -34,16 +30,21 @@ check_armadillo_update() {
   
   if [ "$VERSION_USED" != "$ARMADILLO_LATEST_VERSION" ]; then
   {
-    #Stop Armadillo
-    systemctl stop armadillo
+       
         
     DL_URL=https://github.com/molgenis/molgenis-service-armadillo/releases/download/armadillo-service-$ARMADILLO_LATEST_VERSION/armadillo-$ARMADILLO_LATEST_VERSION.jar
           
-    wget -q -O $ARMADILLO_PATH/application/armadillo-"$ARMADILLO_LATEST_VERSION".jar "$DL_URL"
-    ln -s -f $ARMADILLO_PATH/application/armadillo-"$ARMADILLO_LATEST_VERSION".jar $ARMADILLO_PATH/application/armadillo.jar
-    echo "$ARMADILLO_LATEST_VERSION updated"
-    systemctl start armadillo
+    
+    if validate_url $DL_URL; then
+      # Stop armadillo
+      systemctl stop armadillo
+      wget -q -O $ARMADILLO_PATH/application/armadillo-"$ARMADILLO_LATEST_VERSION".jar "$DL_URL"
+      ln -s -f $ARMADILLO_PATH/application/armadillo-"$ARMADILLO_LATEST_VERSION".jar $ARMADILLO_PATH/application/armadillo.jar
+      echo "$ARMADILLO_LATEST_VERSION updated"
+      systemctl start armadillo
 
+    fi
+    
   }
   else
   {
@@ -51,11 +52,18 @@ check_armadillo_update() {
   }
   fi
 
-
-
   #echo $VER
 
 }
 
+function validate_url(){
+  if [[ `wget -S --spider $1  2>&1 | grep 'HTTP/1.1 200 OK'` ]]; then
+    return 0
+  else
+    return 1
+  fi
+}
 
+
+check_self_update
 check_armadillo_update
